requirements.txt
discord
pony
python-box
